Modified some Dungeon levels to make difficulty more balanced.
- Dungeon bottom-left stage (0,4) changed to make stealth section easier
- Moved introduction of Elite Beanpoles to 3rd room from 2nd room
- On room (2, 4), added additional wall for cover when sneaking around
- Added and fixed some small documentation in a few files
- Closing the game no longer deletes the Save-README.md file

+++ src/helpers/GameState.java
@@ -156,7 +156,8 @@ public class GameState {
 		File tempFolder = new File(GPath.SAVE + "temp");
 		File[] areaFileList = tempFolder.listFiles();
 		for(File areaFile : areaFileList) {
-			areaFile.delete();
+			if(areaFile.getName().endsWith(".ser"))
+				areaFile.delete();
 		}
 	}
